+++ SmartKitchen.Domain/IServices/ICellService.cs
@@ -17,6 +17,6 @@ namespace SmartKitchen.Domain.IServices
         void DeleteCell(Cell cell);
         IQueryable<CellDisplayModel> GetCellsOfStorage(int storageId, string email);
         ServiceResponse UpdateCellBestBefore(int id, DateTime? value, string email);
-        ServiceResponse UpdateCellAmount(int id, decimal value, string email);
+        ServiceResponse UpdateCellAmount(int id, int value, string email);
     }
 }
